Moved factory and updated related code (#6957)
